requirements.txt
python-telegram-bot[job-queue]>=21.0,<22.0
motor>=3.3,<4.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
python-dotenv>=1.0,<2.0
APScheduler>=3.10,<4.0
pytz>=2024.1
structlog>=24.0
pytest>=8.0
pytest-asyncio>=0.23
